move shim from config to utils
