+++ packages/wdio-config/package.json
@@ -31,7 +31,6 @@
   },
   "dependencies": {
     "@wdio/logger": "5.13.2",
-    "@wdio/utils": "^5.13.2",
     "deepmerge": "^4.0.0",
     "glob": "^7.1.2"
   },
