+++ packages/wdio-config/src/index.js
@@ -1,10 +1,5 @@
 import ConfigParser from './lib/ConfigParser'
 import { validateConfig, getSauceEndpoint, detectBackend } from './utils'
-import {
-    wrapCommand, runFnInFiberContext, executeHooksWithArgs,
-    hasWdioSyncSupport, executeSync, executeAsync,
-} from './shim'
-import { testFnWrapper, runTestInFiberContext } from './test-framework'
 import { DEFAULT_CONFIGS } from './constants'
 
 export {
@@ -12,18 +7,6 @@ export {
     getSauceEndpoint,
     detectBackend,
     ConfigParser,
-
-    /**
-     * wdio-sync shim
-     */
-    wrapCommand,
-    executeSync,
-    executeAsync,
-    runFnInFiberContext,
-    runTestInFiberContext,
-    testFnWrapper,
-    executeHooksWithArgs,
-    hasWdioSyncSupport,
 
     /**
      * constants
