+++ packages/wdio-config/src/utils.js
@@ -138,11 +138,3 @@ export function validateConfig (defaults, options) {
 
     return params
 }
-
-/**
- * filter out arguments passed to specFn & hookFn, don't allow callbacks
- * as there is no need for user to call e.g. `done()`
- */
-export function filterSpecArgs (args) {
-    return args.filter((arg) => typeof arg !== 'function')
-}
