+++ packages/wdio-config/tests/__mocks__/@wdio/config.js
@@ -36,16 +36,4 @@ export const validateConfig = jest.fn().mockImplementation(
         config
     )
 )
-export const executeSync = jest.fn()
-export const executeAsync = jest.fn()
-export const wrapCommand = (_, origFn) => origFn
 export const ConfigParser = ConfigParserMock
-export const runTestInFiberContext = jest.fn().mockReturnValue(jest.fn())
-export const executeHooksWithArgs = jest.fn()
-export const runFnInFiberContext = jest.fn().mockImplementation((fn) => {
-    return function (...args) {
-        return Promise.resolve(fn.apply(this, args))
-    }
-})
-export let hasWdioSyncSupport = false
-export const testFnWrapper = jest.fn()
