+++ packages/wdio-cucumber-framework/package.json
@@ -30,7 +30,6 @@
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
-    "@wdio/config": "5.13.2",
     "@wdio/logger": "5.13.2",
     "@wdio/sync": "5.13.2",
     "@wdio/utils": "5.13.2",
