+++ packages/wdio-jasmine-framework/package.json
@@ -36,8 +36,8 @@
     "@types/jasmine": "^3.3.8"
   },
   "dependencies": {
-    "@wdio/config": "5.13.2",
     "@wdio/logger": "5.13.2",
+    "@wdio/utils": "5.13.2",
     "jasmine": "^3.3.0"
   },
   "peerDependencies": {
