+++ packages/wdio-mocha-framework/package.json
@@ -30,8 +30,8 @@
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
-    "@wdio/config": "5.13.2",
     "@wdio/logger": "5.13.2",
+    "@wdio/utils": "5.13.2",
     "mocha": "^6.1.0"
   },
   "devDependencies": {
