+++ packages/wdio-sync/package.json
@@ -34,8 +34,7 @@
     "url": "https://github.com/webdriverio/webdriverio/issues"
   },
   "dependencies": {
-    "@wdio/logger": "5.13.2",
-    "@wdio/utils": "5.13.2"
+    "@wdio/logger": "5.13.2"
   },
   "optionalDependencies": {
     "fibers": "^4.0.1",
