+++ packages/wdio-utils/src/index.js
@@ -2,17 +2,32 @@ import initialisePlugin from './initialisePlugin'
 import initialiseServices from './initialiseServices'
 import webdriverMonad from './monad'
 import {
-    commandCallStructure, isValidParameter, getArgumentType, safeRequire,
-    isFunctionAsync
+    commandCallStructure, isValidParameter, getArgumentType, safeRequire
 } from './utils'
+import {
+    wrapCommand, runFnInFiberContext, executeHooksWithArgs,
+    hasWdioSyncSupport, executeSync, executeAsync,
+} from './shim'
+import { testFnWrapper, runTestInFiberContext } from './test-framework'
 
 export {
     initialisePlugin,
     initialiseServices,
-    isFunctionAsync,
     webdriverMonad,
     commandCallStructure,
     isValidParameter,
     getArgumentType,
-    safeRequire
+    safeRequire,
+
+    /**
+     * wdio-sync shim
+     */
+    wrapCommand,
+    executeSync,
+    executeAsync,
+    runFnInFiberContext,
+    runTestInFiberContext,
+    testFnWrapper,
+    executeHooksWithArgs,
+    hasWdioSyncSupport
 }
