+++ packages/wdio-utils/src/test-framework/errorHandler.js
