+++ packages/wdio-utils/tests/__mocks__/@wdio/utils.js
@@ -1,5 +1,4 @@
 import {
-    isFunctionAsync as isFnAsync,
     getArgumentType as getArgumentTypeOrig,
     isValidParameter as isValidParameterOrig,
     commandCallStructure as commandCallStructureOrig
@@ -46,7 +45,22 @@ export const initialisePlugin = jest.fn().mockImplementation(
 export const initialiseServices = jest.fn().mockReturnValue([])
 export const isValidParameter = isValidParameterOrig
 export const commandCallStructure = commandCallStructureOrig
-export const isFunctionAsync = isFnAsync
 export const safeRequire = jest.fn().mockReturnValue(() => {})
 export const webdriverMonad = webdriverMonadOrig
 export const getArgumentType = getArgumentTypeOrig
+
+/**
+ * shim
+ */
+export const executeSync = jest.fn()
+export const executeAsync = jest.fn()
+export const wrapCommand = (_, origFn) => origFn
+export const runTestInFiberContext = jest.fn().mockReturnValue(jest.fn())
+export const executeHooksWithArgs = jest.fn()
+export const runFnInFiberContext = jest.fn().mockImplementation((fn) => {
+    return function (...args) {
+        return Promise.resolve(fn.apply(this, args))
+    }
+})
+export let hasWdioSyncSupport = false
+export const testFnWrapper = jest.fn()
