+++ packages/wdio-utils/tests/shim-sync.test.js
