+++ packages/wdio-utils/tests/test-framework/errorHandler.test.js
