+++ packages/wdio-utils/tests/test-framework/testFnWrapper-sync.test.js
